+++ src/amg/HISTORY.rst
@@ -10,3 +10,7 @@ Release History
 0.1.1
 ++++++
 * update 'az grafana delete' to automatically remove the default role assignment created for the managed identity
+
+0.1.2
+++++++
+* `az grafana create`: support guess users